Rename ModalTransition to VerticalSheetTransition and add presentation controller support.
Summary:
The new implementation makes use of a presentation controller to allow clients to customize the presented view controller's frame. When a custom frame is provided, the transition will show a dimmed scrim behind the presented view controller that can be tapped in order to dismiss the view controller.

We will likely need to explore APIs for customizing the tap-to-dismiss behavior. For example, modal dialogs that require confirmation may not have a cancelable action and therefor shouldn't be dismissable with a tap on the scrim.

Reviewers: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, randcode-generator, markwei

Reviewed By: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, randcode-generator, markwei

Tags: #material_motion

Differential Revision: http://codereview.cc/D3231