Add a FABFullScreenTransition component.
Summary:
This component makes it possible to present a view controller using a masked reveal transition from a contextual fab button.

The presenting view controller is expected to implement TransitionContextViewRetriever. This isn't enforced at compile time in any way unfortunately, so we may want to explore alternative APIs for ensuring that the context view can be retrieved. Now that we are providing instances to the transition controller it may be reasoanble to simply provide the context view to the transition object. We may still find value in the context view retrieval APIs for app state restoration, but this can be an optional API that an app chooses to opt in to.

Reviewers: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, randcode-generator, markwei

Reviewed By: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, randcode-generator, markwei

Subscribers: markwei

Tags: #material_motion

Differential Revision: http://codereview.cc/D3232