Fine tune the fab transition to support different contexts.
Summary:
Added a new fab transition example to the catalog. This new example allows you to pick the type of target the fab will animate to.

The fab transition now makes use of an internal Motion type that captures the motion information from our fab specifications. The transition can be thought of in terms of three essential parts:

1. The motion information. These are the raw numbers we'll use in different contexts, e.g. toolbar collapse or fullscreen expansion.
2. The motion router. Inspects the transition context and chooses which motion to use.
3. The animations. Turns a motion object into a set of animations and applies them to the relevant properties.

Reviewers: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, randcode-generator

Reviewed By: randcode-generator

Tags: #material_motion

Differential Revision: http://codereview.cc/D3240